HEAD catchup + minor fixes

+++ src/avxsynth/avssources.h
@@ -53,7 +53,7 @@ public:
 		const char *ConvertToFormatName, const char *VarPrefix, IScriptEnvironment* Env);
 	~AvisynthVideoSource();
 	bool __stdcall GetParity(int n);
-	void __stdcall SetCacheHints(int cachehints, size_t frame_range) { }
+	void __stdcall SetCacheHints(int cachehints, int frame_range) { }
 	const VideoInfo& __stdcall GetVideoInfo() { return VI; }
 	void __stdcall GetAudio(void* Buf, __int64 Start, __int64 Count, IScriptEnvironment *Env) { }
 	PVideoFrame __stdcall GetFrame(int n, IScriptEnvironment *Env);
@@ -68,7 +68,7 @@ public:
 		int AdjustDelay, const char *VarPrefix, IScriptEnvironment* Env);
 	~AvisynthAudioSource();
 	bool __stdcall GetParity(int n) { return false; }
-	void __stdcall SetCacheHints(int cachehints, size_t frame_range) { }
+	void __stdcall SetCacheHints(int cachehints, int frame_range) { }
 	const VideoInfo& __stdcall GetVideoInfo() { return VI; }
 	void __stdcall GetAudio(void* Buf, __int64 Start, __int64 Count, IScriptEnvironment *Env);
 	PVideoFrame __stdcall GetFrame(int n, IScriptEnvironment *Env) { return NULL; };
